+++ projects/ngx-form-object/src/lib/form-object-builder/form-object-builder.ts
@@ -1,4 +1,4 @@
-import { FormBuilder, ValidatorFn } from '@angular/forms';
+import { AbstractControl, FormBuilder, ValidatorFn } from '@angular/forms';
 import { ExtendedFormArray } from '../extended-form-array/extended-form-array';
 import { ExtendedFormControl } from '../extended-form-control/extended-form-control';
 import { FormObject } from '../form-object/form-object';
@@ -13,16 +13,16 @@ export class FormObjectBuilder {
 		this.formBuilder = new FormBuilder();
 	}
 
-	public create(formObject: FormObject): FormStore {
-		const formFields = {};
+	public create<T>(formObject: FormObject<T>): FormStore<T> {
+		const formFields: Record<string, AbstractControl> = {};
 
 		Object.assign(formFields, this.createAttributeFormFields(formObject));
 		Object.assign(formFields, this.createHasManyFormFields(formObject));
 		Object.assign(formFields, this.createBelongsToFormFields(formObject));
 
-		const formStoreClass: any = formObject.formStoreClass ? formObject.formStoreClass : FormStore;
+		const formStoreClass = formObject.formStoreClass || FormStore;
 
-		const formStore: FormStore = new formStoreClass(
+		const formStore: FormStore<T> = new formStoreClass(
 			formFields,
 			formObject.formGroupOptions.validator,
 			formObject.formGroupOptions.asyncValidator
@@ -32,10 +32,10 @@ export class FormObjectBuilder {
 		return formStore;
 	}
 
-	private createAttributeFormFields(formObject: FormObject): Record<string, unknown> {
-		const attributeFormFields = {};
+	private createAttributeFormFields<T>(formObject: FormObject<T>): Record<string, unknown> {
+		const attributeFormFields: Record<string, AbstractControl> = {};
 
-		formObject.attributePropertiesKeys.forEach((attributeName: string | symbol) => {
+		formObject.attributePropertiesKeys.forEach((attributeName: string) => {
 			const buildFunction = formObject[`build${capitalize(attributeName.toString())}`];
 			const validators: ValidatorFn | Array<ValidatorFn> = formObject.getValidators(attributeName.toString());
 			const maskFunction: Function = formObject[`mask${capitalize(attributeName.toString())}`];
@@ -52,10 +52,10 @@ export class FormObjectBuilder {
 		return attributeFormFields;
 	}
 
-	private createHasManyFormFields(formObject: FormObject): Record<string, unknown> {
-		const hasManyFormFields = {};
+	private createHasManyFormFields<T>(formObject: FormObject<T>): Record<string, unknown> {
+		const hasManyFormFields: Record<string, AbstractControl> = {};
 
-		formObject.hasManyPropertiesKeys.forEach((propertyName) => {
+		formObject.hasManyPropertiesKeys.forEach((propertyName: string) => {
 			const buildFunction = formObject[`build${capitalize(propertyName.toString())}`];
 			const validators: ValidatorFn | Array<ValidatorFn> = formObject.getValidators(propertyName.toString());
 			const hasManyModels = formObject.model[propertyName];
@@ -69,7 +69,7 @@ export class FormObjectBuilder {
 		return hasManyFormFields;
 	}
 
-	private createBelongsToFormFields(formObject: FormObject): Record<string, unknown> {
+	private createBelongsToFormFields<T>(formObject: FormObject<T>): Record<string, unknown> {
 		const belongsToFormFields = {};
 
 		formObject.belongsToPropertiesKeys.forEach((propertyName: string | symbol) => {
@@ -103,8 +103,8 @@ export class FormObjectBuilder {
 		return belongsToFormFields;
 	}
 
-	private buildRelationshipModels(
-		formObject: FormObject,
+	private buildRelationshipModels<T>(
+		formObject: FormObject<T>,
 		relationshipName: string | symbol,
 		relationshipModels: Array<any> = [],
 		propertyOptions: PropertyOptions
@@ -113,7 +113,11 @@ export class FormObjectBuilder {
 		const formGroups: Array<any> = [];
 
 		relationshipModels.forEach((relationshipModel) => {
-			const formStore: FormStore = this.createRelationshipFormObject(formObject, relationshipName, relationshipModel);
+			const formStore: FormStore<T> = this.createRelationshipFormObject(
+				formObject,
+				relationshipName,
+				relationshipModel
+			);
 			if (formStore) {
 				formGroups.push(formStore);
 			}
@@ -129,22 +133,22 @@ export class FormObjectBuilder {
 		return relationshipFormGroups;
 	}
 
-	private createRelationshipFormObject(
-		formObject: FormObject,
+	private createRelationshipFormObject<T>(
+		formObject: FormObject<T>,
 		relationshipName: string | symbol,
 		relationshipModel: any,
 		propertyOptions: PropertyOptions = {}
-	): FormStore {
+	): FormStore<T> {
 		const createFormObjectFunction = formObject[`create${capitalize(relationshipName.toString())}FormObject`];
 
 		if (createFormObjectFunction) {
-			const modelFormObject: FormObject = createFormObjectFunction.call(
+			const modelFormObject: FormObject<T> = createFormObjectFunction.call(
 				formObject,
 				relationshipModel,
 				null,
 				propertyOptions
 			);
-			const formStore: FormStore = this.create(modelFormObject);
+			const formStore: FormStore<T> = this.create(modelFormObject);
 			return formStore;
 		} else {
 			// There is no function specified for creating form object for ${relationshipName.toString()}