+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### 9.0.4-beta
+
+  * [BREAKING CHANGE] `FormStore` uses a generic now
+  * [BREAKING CHANGE] `FormObject` uses a generic now
+
 ### 9.0.3-beta
 
   * Temporary adjust save method types